+++ src/seacoral-c/cmd.ml
@@ -74,6 +74,13 @@ module Log_lwt_clang = (val Ez_logs.subproc "clang")
 module Log_lwt_clangxx = (val Ez_logs.subproc "clang++")
 module Log_lwt_ld = (val Ez_logs.subproc "ld")
 
+let stream_grab_xor_log grabber logger = match grabber with
+  | None -> `Log logger
+  | Some grabber ->
+     (* S: Using `GrabNLog with dummy logger because using `Grab
+        leads to an infinite loop. *)
+     `GrabNLog (grabber, (fun ?header:_ _ _ -> Lwt.return ()))
+
 let stream_grab grabber logger = match grabber with
   | None -> `Log logger
   | Some grabber -> `GrabNLog (grabber, logger)
@@ -122,6 +129,29 @@ let clang_check
     ~stdout:(stream_grab stdout_grabber Log_lwt_clang.LWT.debug)
     ~stderr:(stream_grab stderr_grabber Log_lwt_clang.LWT.debug)
     ~on_error:(cc_error Syntax_check_file file_name)
+
+let clang_check_and_print_llvm
+    ?clang_cmd
+    ?(cflags = ["-g"])
+    ?(cppflags = [])
+    ?stdout_grabber ?stderr_grabber
+    (file: [< `C | `CXX] Sc_sys.File.t)
+  : unit Lwt.t
+  =
+  let file_name = Sc_sys.File.name file in
+  Sc_sys.Process.PRETTY.shell_unit
+    (* TODO: -fcolor-diagnostics only if at least one log reporter accepts TTY
+       control characters. *)
+    "%s %a%a%a-c %s -fsyntax-only -fcolor-diagnostics -Xclang -ast-dump"
+    (lazy_cmd ENV.clang_exe clang_cmd)
+    Basics.PPrt.Strings.pp_space_separated_ cflags
+    Basics.PPrt.Strings.pp_space_separated_ cppflags
+    Fmt.(option @@ fmt "%s ") (Lazy.force ENV.cppflags)
+    file_name
+    ~stdout:(stream_grab_xor_log stdout_grabber Log_lwt_clang.LWT.debug) 
+    ~stderr:(stream_grab stderr_grabber Log_lwt_clang.LWT.debug)
+    ~on_error:(cc_error Syntax_check_file file_name)
+
 
 (** {2 Compiling C/C++ source code} *)